% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJsonRobject.R
\name{convertJsonRobject}
\alias{convertJsonRobject}
\title{Transform JsonMixtComp output}
\usage{
convertJsonRobject(out, confidenceLevel = 0.95, mode = c("learn",
  "predict"))
}
\arguments{
\item{out}{output of JsonMixtComp}

\item{confidenceLevel}{quantile for confidence interval of estimated parameters.}

\item{mode}{"learn" or "predict"}
}
\value{
list that looks like le output of RMixtCompPredict and RMixtCompCluster
}
\description{
Transform JsonMixtComp output
}
