% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonMixtComp.R
\name{createJsonMixtCompFile}
\alias{createJsonMixtCompFile}
\title{Create a json file for JsonMixtComp}
\usage{
createJsonMixtCompFile(outputJsonFile, dataList, nbClass,
  confidenceLevel = 0.95, mcStrategy = list(nbBurnInIter = 100, nbIter =
  100, nbGibbsBurnInIter = 100, nbGibbsIter = 100, parameterEdgeAuthorized =
  TRUE), mode = c("learn", "predict"), outMixtCompFile = NULL)
}
\arguments{
\item{outputJsonFile}{\emph{lm} output of \link{getData} function}

\item{dataList}{output of getData function.}

\item{nbClass}{the number of class of the mixture model.}

\item{confidenceLevel}{quantile for confidence interval of estimated parameters.}

\item{mcStrategy}{a list containing the parameters of the SEM-Gibbs algorithm (see \emph{Details}).}

\item{mode}{"learn" or "predict". Mode used in JsonMixtComp.}

\item{outMixtCompFile}{(only if mode = "predict"). Output json file of JsonMixtComp in learn mode.}
}
\value{
a json file to use with JsonMixtComp.
}
\description{
Create a json file containing the data and options required by JsonMixtComp.
}
