% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_misc.R
\name{createMcStrategy}
\alias{createMcStrategy}
\title{create mcStrategy object}
\usage{
createMcStrategy(nbBurnInIter = 100, nbIter = 100, nbGibbsBurnInIter = 50,
  nbGibbsIter = 50, nInitPerClass = 10, nSemTry = 20)
}
\arguments{
\item{nbBurnInIter}{Number of iterations of the burn-in part of the SEM algorithm.}

\item{nbIter}{Number of iterations of the SEM algorithm.}

\item{nbGibbsBurnInIter}{Number of iterations of the burn-in part of the Gibbs algorithm.}

\item{nbGibbsIter}{Number of iterations of the Gibbs algorithm.}

\item{nInitPerClass}{Number of individuals used to initialize each cluster (default = 10).}

\item{nSemTry}{Number of try of the algorithm for avoiding an error.}
}
\value{
a list with the parameters values
}
\description{
create a mcStrategy object
}
\examples{
# default values
mcStrategy <- createMcStrategy()

# default values
mcStrategy <- createMcStrategy(nbIter = 200)

}
