% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getBIC}
\alias{getBIC}
\alias{getICL}
\title{Get criterion value}
\usage{
getBIC(outMixtComp)
getICL(outMixtComp)
}
\arguments{
\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict} functions.}
}
\value{
value of the criterion
}
\description{
Get criterion value
}
