% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getData}
\alias{getData}
\title{Read data and descriptor files}
\usage{
getData(...)
}
\arguments{
\item{...}{a vector of two elements : the path to the data file and the path to the descriptor file.}
}
\value{
a list
}
\description{
Read csv file of data and descriptor and create an R object for \link{JsonMixtCompCluster} and \link{JsonMixtCompPredict}.
}
\examples{
\dontrun{
pathToData <- system.file("extdata", "data.csv", package = "RJsonMixtComp")
pathToDescriptor <- system.file("extdata", "descriptor.csv", package = "RJsonMixtComp")

res <- getData(c(pathToData, pathToDescriptor))#' 
}

}
\seealso{
\link{JsonMixtCompCluster} \link{JsonMixtCompPredict}
}
