% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getEmpiricTik}
\alias{getEmpiricTik}
\alias{getTik}
\title{Get the tik}
\usage{
getEmpiricTik(outMixtComp)

getTik(outMixtComp, log = TRUE)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}

\item{log}{if TRUE, log(tik) are returned}
}
\value{
a matrix containing the tik for each individuals (in row) and each class (in column).
}
\description{
Get the a posteriori probability to belong to each class for each individual
}
\details{
\emph{getTik} returns a posteriori probabilities computed with the returned parameters. \emph{getEmpiricTik} returns an estimation based on the sampled z_i during the algorithm.
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2)

# get tik
tikEmp <- getEmpiricTik(resLearn)
tik <- getTik(resLearn, log = FALSE)
}

}
