% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getParam}
\alias{getParam}
\title{Get the estimated parameter}
\usage{
getParam(varName, outMixtComp)
}
\arguments{
\item{varName}{name of the variable to get parameter}

\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict}  functions.}
}
\value{
the parameter of the variable
}
\description{
Get the estimated parameter
}
