% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getTik}
\alias{getTik}
\title{Get the empiric tik}
\usage{
getTik(outMixtComp)
}
\arguments{
\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict} functions.}
}
\value{
a matrix containing the tik for each individuals (in row) and each class (in column).
}
\description{
Get the empiric tik
}
