% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getType}
\alias{getType}
\title{Get the type of model for each variable}
\usage{
getType(outMixtComp, with.z_class = FALSE)
}
\arguments{
\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict} functions.}

\item{with.z_class}{if TRUE, the type of z_class is returned.}
}
\value{
a vector containing the type of models associated with each individual.
}
\description{
Get the type of model for each variable
}
