% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{getZ_class}
\alias{getZ_class}
\title{Get the estimated class from MixtComp object}
\usage{
getZ_class(outMixtComp)
}
\arguments{
\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict} functions.}
}
\value{
a vector containing the estimated class for each individual.
}
\description{
Get the estimated class from MixtComp object
}
