% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonMixtComp.R
\name{JsonMixtCompCluster}
\alias{JsonMixtCompCluster}
\alias{JsonMixtCompPredict}
\title{Learn and predict using RJsonMixtComp}
\usage{
JsonMixtCompCluster(dataList, mcStrategy, nbClass, confidenceLevel, jsonInputFile,
                    jsonOutputFile)
JsonMixtCompPredict(dataList, mcStrategy, nbClass, confidenceLevel, jsonInputFile, 
                    jsonOutputFile, jsonMixtCompLearnFile)
}
\arguments{
\item{dataList}{\emph{lm} output of \link{getData} function.}

\item{mcStrategy}{a list containing the parameters of the SEM-Gibbs algorithm (see \emph{Details}).}

\item{nbClass}{the number of class of the mixture model.}

\item{confidenceLevel}{quantile for confidence interval of estimated parameters.}

\item{jsonInputFile}{path of the input json file to save.}

\item{jsonOutputFile}{path of the output json file to save.}

\item{jsonMixtCompLearnFile}{(only for JsonMixtCompPredict) path of the output json file of JsonMixtCompLearn function.}
}
\value{
A json file and a R list containing 3 lists :
\describe{
 \item{strategy}{a copy of \emph{mcStrategy} parameter.}
 \item{mixture}{information about the mixture (see \emph{Details}).}
 \item{variable}{information about the estimated parameters (see \emph{Details}).}
}
}
\description{
Estimate the parameter of a mixture model or predict the cluster of new samples.
}
\details{
Details about the output object of \emph{mixtCompCluster} and \emph{mixtCompPredict} functions.

The path for outputs files must not contain ":" or "~".

Rank data and ordinal data are currently not working.
}
\examples{
\dontrun{
# get the path to the data of the package
pathToData <- system.file("extdata", "data.csv", package = "RJsonMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RJsonMixtComp")

resGetData <- RJsonMixtComp:::getData(c(pathToData, pathToDescriptor)) 

# learn
resLearn <- JsonMixtCompCluster(dataList = resGetData$lm,
                                mcStrategy = list(nbBurnInIter = 100, nbIter = 100, 
                                                  nbGibbsBurnInIter = 100, nbGibbsIter = 100),
                                nbClass = 2, confidenceLevel = 0.95, 
                                jsonInputFile = "datalearn.json",
                                jsonOutputFile = "reslearn.json")



# predict : require a json file output from JsonMixtCompCluster ("reslearn.json" here)
resPredict <- JsonMixtCompPredict(dataList = resGetData$lm,
                                  mcStrategy = list(nbBurnInIter = 100, nbIter = 100, 
                                                    nbGibbsBurnInIter = 100, nbGibbsIter = 100),
                                  nbClass = 2, confidenceLevel = 0.95, 
                                  jsonInputFile = "datalearn.json",
                                  jsonOutputFile = "respredict.json",
                                  jsonMixtCompLearnFile = "reslearn.json")

# remove created files of the example
file.remove(c("reslearn.json", "respredict.json", "datalearn.json"))
}


}
\seealso{
\link{getData}
}
