% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompUtilities.R
\name{writeJsonParamFile}
\alias{writeJsonParamFile}
\title{Convert an R object to json}
\usage{
writeJsonParamFile(outMixtComp, filename)
}
\arguments{
\item{outMixtComp}{output object of \link{JsonMixtCompCluster} or \link{JsonMixtCompPredict}  functions.}

\item{filename}{name of the variable to get parameter}
}
\description{
convert an R object to a json file to use with \link{JsonMixtCompPredict} function
}
