% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getType}
\alias{getType}
\alias{getVarNames}
\title{Getters}
\usage{
getType(outMixtComp, with.z_class = FALSE)

getVarNames(outMixtComp, with.z_class = FALSE)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} functions.}

\item{with.z_class}{if TRUE, the type of z_class is returned.}
}
\value{
a vector containing the type of models, names associated with each individual.
}
\description{
Get the type of model, names for each variable
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get type
type <- getType(res)

# get variable names
varNames <- getVarNames(res)

}

}
