% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDescription.R
\docType{package}
\name{RJsonMixtComp-package}
\alias{RJsonMixtComp-package}
\title{RJsonMixtComp}
\description{
Statistics analysis with MixtComp
}
\details{
\tabular{ll}{
Package: \tab RJsonMixtComp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-02-22\cr
License: \tab  \cr
}
}
\examples{
\dontrun{
# Do Stuff
}


}
\author{
Your Name Here

Maintainer: Your Name Here <YourNameHere@inria.fr>
}
\keyword{package}
