% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDescription.R
\docType{package}
\name{RJsonMixtComp-package}
\alias{RJsonMixtComp-package}
\title{RJsonMixtComp}
\description{
Statistics analysis with MixtComp
}
\details{
Use JsonMixtComp
}
\examples{
\donttest{
pathToData <- system.file("extdata", "data.json", package = "RJsonMixtComp")
pathToDescriptor <- system.file("extdata", "desc.json", package = "RJsonMixtComp")

data <- as.data.frame(fromJSON(pathToData))
descriptor <- as.data.frame(lapply(fromJSON(pathToDescriptor), unlist))
strategy <- list(nbBurnInIter = 50, nbIter = 50, nbGibbsBurnInIter = 20,
                 nbGibbsIter = 20, nInitPerClass = 10, nSemTry = 5)

resLearn <- JsonMixtCompLearn(data, descriptor, nClass = 2, mcStrategy = strategy,
                              confidenceLevel = 0.95, inputPath = ".", outputFile = "reslearn.json")


file.remove("./algo.json", "./descriptor.json", "./data.json")

resPredict <- JsonMixtCompPredict(data, descriptor, nClass = 2, mcStrategy = strategy,
                                  confidenceLevel = 0.95, inputPath = ".", 
                                  paramFile = "reslearn.json", outputFile = "respredict.json")


file.remove("./algo.json", "./descriptor.json", "./data.json", "reslearn.json",
            "respredict.json", "progress")
}


}
\author{
Quentin Grimonprez

Maintainer: Quentin Grimonprez <quentin.grimonprez@inria.fr>
}
\keyword{package}
